refactor: split the two files that crossed the 800-line ceiling
Replaces the allowlist entries added a commit ago. The ceiling is a DoD gate and
it is there for a reason; allowlisting was the easy road and the wrong one.

internal/providers/auth.go 849 -> 469, with the browser-cookie fallback and the
interactive escape hatch moved to auth_browser.go (403). That is the half that
grew, and it is the half with its own threat model: it decides when trvl may
read a user's browser cookies, which destinations those cookies may be sent to,
and when an interactive browser may be opened at all. In its own file that
surface is reviewable on its own terms rather than as a section of a longer file
about request preflight.

mcp/tools_providers_test.go 822 -> 540, with schema, parsing and
tool-definition tests moved to tools_providers_schema_test.go (298). The seam is
subject, not size: everything moved asserts a declared surface or a pure
function and drives no handler. Elicitation, credential redaction and
concurrency stay together.

Both moves are verbatim -- functions travel with their doc comments and no body
was edited. A refactor that also changes code cannot be reviewed as a refactor,
and one of these files is provider authentication.

One test needed a real change, not a move: TestRecoveryTiersDoNotReferenceAuthMu
read auth.go by name to check that no recovery tier acquires authMu, and three
of the four tiers moved. Its own failure message asked for the guard to move
with them. It now scans every non-test file in the package via a new
packageSource helper -- the same traversal selfLockingHelpers already used --
so it survives the next split too. That is stronger than before: the guard is
tied to the property rather than to a filename.

Verification: internal/providers 921 passed, mcp 1224 passed, go build clean,
go vet clean, gofmt clean, and all four scripts/ci hygiene checks pass with no
allowlist entry for either file.

trvl#560 can close once this reaches origin/main.
